SOLID восстановлен
